+++ src/dcos_e2e_cli/dcos_aws/commands/doctor.py
@@ -2,11 +2,9 @@
 Checks for showing up common sources of errors with the AWS backend.
 """
 
-import sys
-
 import click
 
-from dcos_e2e_cli.common.doctor import CheckLevels, check_ssh
+from dcos_e2e_cli.common.doctor import check_ssh, run_doctor_commands
 from dcos_e2e_cli.common.options import verbosity_option
 from dcos_e2e_cli.common.utils import set_logging
 
@@ -19,7 +17,4 @@ def doctor(verbose: int) -> None:
     """
     set_logging(verbosity_level=verbose)
     check_functions = [check_ssh]
-    for function in check_functions:
-        level = function()
-        if level == CheckLevels.ERROR:
-            sys.exit(1)
+    run_doctor_commands(check_functions=check_functions)