+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Next
 ----
 
 - Backwards incompatible change: Changed CLI commands from ``dcos-docker`` to ``minidcos docker`` and alike.
+- Add a ``dcos-docker doctor`` check for systemd.
 
 2018.11.09.1
 ------------